+++ src/main/java/Backend/BuiltinExpressions/MandelExpression.java
@@ -12,19 +12,19 @@ public class MandelExpression extends FunctionExpression {
     }
 
     @Override
-    public double evaluate(Map<String, Double> arguments) {
-        double cx = getInputs()[0].evaluate(arguments);
-        double cy = getInputs()[1].evaluate(arguments);
+    public float evaluate(Map<String, Float> arguments) {
+        float cx = getInputs()[0].evaluate(arguments);
+        float cy = getInputs()[1].evaluate(arguments);
 
-        double x = 0;
-        double y = 0;
+        float x = 0;
+        float y = 0;
         int i;
         for (i = 0; i < 100; i++) {
             if (x*x + y*y > 4) break;
-            double xtemp = x*x - y*y + cx;
+            float xtemp = x*x - y*y + cx;
             y = 2*x*y + cy;
             x = xtemp;
         }
-        return (double) i / 100.f;
+        return i / 100.f;
     }
 }
